Implement (wire up) HPV vaccination widget and functionality into the UG HPV android client
