+++ opensrp-uganda-hpv/src/main/java/org/smartregister/ug/hpv/activity/BaseActivity.java
@@ -1,5 +1,6 @@
 package org.smartregister.ug.hpv.activity;
 
+import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
 import android.support.v7.app.AlertDialog;
@@ -11,8 +12,11 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import org.apache.commons.lang3.StringUtils;
 import org.smartregister.Context;
 import org.smartregister.configurableviews.ConfigurableViewsLibrary;
+import org.smartregister.immunization.domain.VaccineWrapper;
+import org.smartregister.immunization.listener.VaccinationActionListener;
 import org.smartregister.ug.hpv.R;
 import org.smartregister.ug.hpv.event.LanguageConfigurationEvent;
 import org.smartregister.ug.hpv.util.Utils;
@@ -20,6 +24,7 @@ import org.smartregister.util.Log;
 import org.smartregister.view.activity.DrishtiApplication;
 import org.smartregister.view.activity.SecuredActivity;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -27,13 +32,15 @@ import java.util.Map;
  * Created by ndegwamartin on 09/10/2017.
  */
 
-public abstract class BaseActivity extends SecuredActivity {
+public abstract class BaseActivity extends SecuredActivity{
     private static final int MINIUM_LANG_COUNT = 2;
     protected Toolbar toolbar;
+    private ProgressDialog progressDialog;
 
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        initializeProgressDialog();
     }
 
     @Override
@@ -127,4 +134,34 @@ public abstract class BaseActivity extends SecuredActivity {
         //Overrides
     }
 
+    private void initializeProgressDialog() {
+        progressDialog = new ProgressDialog(this);
+        progressDialog.setCancelable(false);
+        progressDialog.setTitle(getString(R.string.saving_dialog_title));
+        progressDialog.setMessage(getString(R.string.please_wait_message));
+    }
+
+
+    public void showProgressDialog(String title, String message) {
+        if (progressDialog != null) {
+            if (StringUtils.isNotBlank(title)) {
+                progressDialog.setTitle(title);
+            }
+
+            if (StringUtils.isNotBlank(message)) {
+                progressDialog.setMessage(message);
+            }
+            progressDialog.show();
+        }
+    }
+
+    public void showProgressDialog() {
+        showProgressDialog(getString(R.string.saving_dialog_title), getString(R.string.please_wait_message));
+    }
+
+    public void hideProgressDialog() {
+        if (progressDialog != null) {
+            progressDialog.dismiss();
+        }
+    }
 }
