+++ opensrp-uganda-hpv/src/main/java/org/smartregister/ug/hpv/activity/BasePatientDetailActivity.java
@@ -2,43 +2,73 @@ package org.smartregister.ug.hpv.activity;
 
 import android.content.Intent;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Environment;
+import android.os.Handler;
+import android.os.Looper;
 import android.provider.MediaStore;
 import android.support.v4.app.Fragment;
 import android.util.Log;
+import android.util.Pair;
+import android.view.View;
 
+import org.apache.commons.lang3.StringUtils;
+import org.joda.time.DateTime;
 import org.smartregister.commonregistry.CommonPersonObjectClient;
+import org.smartregister.domain.Alert;
+import org.smartregister.immunization.domain.Vaccine;
+import org.smartregister.immunization.domain.VaccineSchedule;
+import org.smartregister.immunization.domain.VaccineWrapper;
+import org.smartregister.immunization.listener.VaccinationActionListener;
+import org.smartregister.immunization.repository.VaccineRepository;
+import org.smartregister.immunization.util.VaccinatorUtils;
+import org.smartregister.immunization.view.VaccineGroup;
 import org.smartregister.ug.hpv.R;
 import org.smartregister.ug.hpv.adapter.HPVRegisterActivityPagerAdapter;
+import org.smartregister.ug.hpv.application.HpvApplication;
 import org.smartregister.ug.hpv.fragment.BasePatientDetailsFragment;
+import org.smartregister.ug.hpv.helper.LocationHelper;
+import org.smartregister.ug.hpv.helper.VaccinationHelper;
 import org.smartregister.ug.hpv.util.Constants;
+import org.smartregister.ug.hpv.view.LocationPickerView;
+import org.smartregister.util.Utils;
 import org.smartregister.view.viewpager.OpenSRPViewPager;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.Serializable;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
+import util.UgandaHpvConstants;
+
+import static org.smartregister.ug.hpv.activity.LoginActivity.getOpenSRPContext;
 
 /**
  * Created by ndegwamartin on 17/11/2017.
  */
 
-public abstract class BasePatientDetailActivity extends BaseActivity {
+public abstract class BasePatientDetailActivity extends BaseActivity implements VaccinationActionListener {
     private static final String TAG = BasePatientDetailActivity.class.getCanonicalName();
     protected Map<String, String> patientDedtails;
     protected File currentfile;
     private static final int REQUEST_TAKE_PHOTO = 1;
+    private ArrayList<VaccineGroup> vaccineGroups;
+    private boolean isChildActive = false;
 
     @Bind(R.id.view_pager)
     protected OpenSRPViewPager mPager;
+
     private CommonPersonObjectClient commonPersonObjectClient;
     Fragment mBaseFragment;
+    private LocationPickerView locationPickerView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,6 +78,8 @@ public abstract class BasePatientDetailActivity extends BaseActivity {
 
         commonPersonObjectClient = (CommonPersonObjectClient) getIntent().getSerializableExtra(Constants.INTENT_KEY.CLIENT_OBJECT);
 
+        locationPickerView = (LocationPickerView) findViewById(R.id.facility_selection);
+        vaccineGroups = new ArrayList<>();
 
         Fragment[] otherFragments = {};
 
@@ -61,7 +93,25 @@ public abstract class BasePatientDetailActivity extends BaseActivity {
 
     }
 
+    @Override
+    protected void onRestoreInstanceState(Bundle savedInstanceState) {
+        super.onRestoreInstanceState(savedInstanceState);
+        Serializable serializable = savedInstanceState.getSerializable(Constants.INTENT_KEY.CLIENT_OBJECT);
+        if (serializable != null && serializable instanceof CommonPersonObjectClient) {
+            commonPersonObjectClient = (CommonPersonObjectClient) serializable;
+        }
+    }
+
     protected abstract Fragment getDetailFragment();
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        if (vaccineGroups != null) {
+            // TODO: might need to add some logic here (similar to Zeir)
+            vaccineGroups = null;
+        }
+    }
 
     @Override
     public boolean onSupportNavigateUp() {
@@ -126,4 +176,245 @@ public abstract class BasePatientDetailActivity extends BaseActivity {
 
     }
 
+
+    @Override
+    public void onVaccinateToday(ArrayList<VaccineWrapper> tags, View v) {
+        if (tags != null && !tags.isEmpty()) {
+            View view = getLastOpenedView();
+            saveVaccine(tags, view);
+        }
+    }
+
+    @Override
+    public void onVaccinateEarlier(ArrayList<VaccineWrapper> tags, View v) {
+        if (tags != null && !tags.isEmpty()) {
+            View view = getLastOpenedView();
+            saveVaccine(tags, view);
+        }
+    }
+
+    @Override
+    public void onUndoVaccination(VaccineWrapper tag, View v) {
+        org.smartregister.util.Utils.startAsyncTask(new UndoVaccineTask(tag, v), null);
+    }
+
+    private void saveVaccine(ArrayList<VaccineWrapper> tags, final View view){
+        if (tags.isEmpty()) {
+            return;
+        }
+
+        VaccineRepository vaccineRepository = HpvApplication.getInstance().vaccineRepository();
+
+        VaccineWrapper[] arrayTags = tags.toArray(new VaccineWrapper[tags.size()]);
+        SaveVaccinesTask backgroundTask = new SaveVaccinesTask();
+        backgroundTask.setVaccineRepository(vaccineRepository);
+        backgroundTask.setView(view);
+        org.smartregister.util.Utils.startAsyncTask(backgroundTask, arrayTags);
+    }
+
+
+    private void saveVaccine(VaccineRepository vaccineRepository, VaccineWrapper tag) {
+        if (tag.getUpdatedVaccineDate() == null) {
+            return;
+        }
+
+        Vaccine vaccine = new Vaccine();
+        if (tag.getDbKey() != null) {
+            vaccine = vaccineRepository.find(tag.getDbKey());
+        }
+        vaccine.setBaseEntityId(commonPersonObjectClient.entityId());
+        vaccine.setName(tag.getName());
+        vaccine.setDate(tag.getUpdatedVaccineDate().toDate());
+        vaccine.setAnmId(getOpenSRPContext().allSharedPreferences().fetchRegisteredANM());
+
+        vaccine.setLocationId(LocationHelper.getInstance().getOpenMrsLocationId(locationPickerView.getSelectedItem()));
+
+        String lastChar = vaccine.getName().substring(vaccine.getName().length() - 1);
+        if (StringUtils.isNumeric(lastChar)) {
+            vaccine.setCalculation(Integer.valueOf(lastChar));
+        } else {
+            vaccine.setCalculation(-1);
+        }
+        org.smartregister.ug.hpv.util.Utils.addVaccine(vaccineRepository, vaccine);
+        tag.setDbKey(vaccine.getId());
+    }
+
+
+    private void addVaccineGroup(int canvasId, org.smartregister.immunization.domain.jsonmapping.VaccineGroup vaccineGroupData, List<Vaccine> vaccineList, List<Alert> alerts) {
+        // TODO: Add logic to add vaccine group to view (similar to Zeir)
+        VaccineGroup curGroup = new VaccineGroup(this);
+        vaccineGroups.add(curGroup);
+    }
+
+
+    private void updateVaccinationViews(List<Vaccine> vaccineList, List<Alert> alerts) {
+        if (vaccineGroups == null) {
+            vaccineGroups = new ArrayList<>();
+            List<org.smartregister.immunization.domain.jsonmapping.VaccineGroup> supportedVaccines = VaccinatorUtils.getSupportedVaccines(this);
+            for (org.smartregister.immunization.domain.jsonmapping.VaccineGroup vaccineGroup : supportedVaccines) {
+                addVaccineGroup(-1, vaccineGroup, vaccineList, alerts);
+            }
+        } else {
+            for(VaccineGroup vaccineGroup : vaccineGroups) {
+                vaccineGroup.setChildActive(isChildActive);
+                vaccineGroup.updateChildsActiveStatus();
+            }
+        }
+    }
+
+
+    private void updateVaccineGroupViews(View view, final ArrayList<VaccineWrapper> wrappers, List<Vaccine> vaccineList) {
+        updateVaccineGroupViews(view, wrappers, vaccineList, false);
+    }
+
+    private void updateVaccineGroupViews(View view, final ArrayList<VaccineWrapper> wrappers, final List<Vaccine> vaccineList, final boolean undo) {
+        if (view == null || !(view instanceof VaccineGroup)) {
+            return;
+        }
+        final VaccineGroup vaccineGroup = (VaccineGroup) view;
+        vaccineGroup.setModalOpen(false);
+
+        if (Looper.myLooper() == Looper.getMainLooper()) {
+            if (undo) {
+                vaccineGroup.setVaccineList(vaccineList);
+                vaccineGroup.updateWrapperStatus(wrappers, UgandaHpvConstants.KEY.CHILD);
+            }
+            vaccineGroup.updateViews(wrappers);
+
+        } else {
+            Handler handler = new Handler(Looper.getMainLooper());
+            handler.post(new Runnable() {
+                @Override
+                public void run() {
+                    if (undo) {
+                        vaccineGroup.setVaccineList(vaccineList);
+                        vaccineGroup.updateWrapperStatus(wrappers, UgandaHpvConstants.KEY.CHILD);
+                    }
+                    vaccineGroup.updateViews(wrappers);
+                }
+            });
+        }
+    }
+
+    private VaccineGroup getLastOpenedView() {
+        if (vaccineGroups == null) {
+            return null;
+        }
+
+        for (VaccineGroup vaccineGroup : vaccineGroups) {
+            if (vaccineGroup.isModalOpen()) {
+                return vaccineGroup;
+            }
+        }
+        return null;
+    }
+
+
+
+
+    //////////////////////////////// AsyncTasks ////////////////////////////////////
+
+
+
+    private class UndoVaccineTask extends AsyncTask<Void, Void, Void> {
+        private final VaccineWrapper tag;
+        private final View view;
+        private final VaccineRepository vaccineRepository;
+        private List<Vaccine> vaccineList;
+        private List<String> affectedVaccines;
+
+        public UndoVaccineTask(VaccineWrapper tag, View view) {
+            this.tag = tag;
+            this.view = view;
+            vaccineRepository = HpvApplication.getInstance().vaccineRepository();
+        }
+
+        @Override
+        protected  void onPreExecute() {
+            showProgressDialog(getString(R.string.updating_dialog_title), null);
+        }
+
+        @Override
+        protected Void doInBackground(Void... params) {
+            if  (tag != null && tag.getDbKey() != null) {
+                Long dbKey = tag.getDbKey();
+                vaccineRepository.deleteVaccine(dbKey);
+
+                String dobString = Utils.getValue(commonPersonObjectClient.getColumnmaps(), UgandaHpvConstants.DOB, false);
+                DateTime dateTime = org.smartregister.ug.hpv.util.Utils.dobStringToDateTime(dobString);
+                if (dateTime != null) {
+                    affectedVaccines = VaccineSchedule.updateOfflineAlerts(commonPersonObjectClient.entityId(), dateTime, UgandaHpvConstants.KEY.CHILD);
+                }
+                vaccineList = vaccineRepository.findByEntityId(commonPersonObjectClient.entityId());
+            }
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(Void params) {
+            hideProgressDialog();
+            super.onPostExecute(params);
+
+            // Refresh the vaccine group with the updated vaccine
+            tag.setUpdatedVaccineDate(null, false);
+            tag.setDbKey(null);
+
+            View view = getLastOpenedView();
+
+            ArrayList<VaccineWrapper> wrappers = new ArrayList<>();
+            wrappers.add(tag);
+            updateVaccineGroupViews(view, wrappers, vaccineList, true);
+        }
+    }
+
+    private class SaveVaccinesTask extends AsyncTask<VaccineWrapper, Void, Pair<ArrayList<VaccineWrapper>, List<Vaccine>>> {
+
+        private View view;
+        private VaccineRepository vaccineRepository;
+        private List<String> affectedVaccines;
+        private List<Vaccine> vaccineList;
+
+        public void setView(View view) {
+            this.view = view;
+        }
+
+        public void setVaccineRepository(VaccineRepository vaccineRepository) {
+            this.vaccineRepository = vaccineRepository;
+            affectedVaccines = new ArrayList<>();
+        }
+
+        @Override
+        protected void onPreExecute() {
+            showProgressDialog();
+        }
+
+        @Override
+        protected void onPostExecute(Pair<ArrayList<VaccineWrapper>, List<Vaccine>> pair) {
+            hideProgressDialog();
+            updateVaccineGroupViews(view, pair.first, pair.second);
+            // TODO: maybe add the show weight logic as in zeir
+        }
+
+        @Override
+        protected Pair<ArrayList<VaccineWrapper>, List<Vaccine>> doInBackground(VaccineWrapper... vaccineWrappers) {
+
+            ArrayList<VaccineWrapper> list = new ArrayList<>();
+            if (vaccineRepository != null) {
+                for (VaccineWrapper tag : vaccineWrappers) {
+                    saveVaccine(vaccineRepository, tag);
+                    list.add(tag);
+                }
+            }
+
+            Pair<ArrayList<VaccineWrapper>, List<Vaccine>> pair = new Pair<>(list, vaccineList);
+            String dobString = org.smartregister.util.Utils.getValue(commonPersonObjectClient.getColumnmaps(), UgandaHpvConstants.DOB, false);
+            DateTime dateTime = org.smartregister.ug.hpv.util.Utils.dobStringToDateTime(dobString);
+            if (dateTime != null) {
+                affectedVaccines = VaccineSchedule.updateOfflineAlerts(commonPersonObjectClient.entityId(), dateTime, UgandaHpvConstants.KEY.CHILD);
+            }
+
+            vaccineList = vaccineRepository.findByEntityId(commonPersonObjectClient.entityId());
+            return pair;
+        }
+    }
 }
